change quantity to long
git-svn-id: https://repo.algotrader.ch/svn/trunk/AlgoTrader@2795 19eef36c-86f0-de11-afab-0004ac968479

+++ src/com/algoTrader/entity/SecurityImpl.java
@@ -33,7 +33,7 @@ public class SecurityImpl extends com.algoTrader.entity.Security {
         return getPosition().getQuantity() > 0;
     }
 
-    public BigDecimal getCommission(int quantity, TransactionType transactionType) {
+    public BigDecimal getCommission(long quantity, TransactionType transactionType) {
 
         return new BigDecimal(0);
     }
